# openapi.yaml
openapi: 3.0.3
info:
  title: Bierliste Backend API
  version: 0.2.0
  description: |
    REST API for the Kickerliste office foosball score tracker.

    ## Authentication
    Most endpoints require a JWT passed as `Authorization: Bearer <token>`.
    Tokens are obtained from `POST /auth/login` or `POST /auth/register`.

    Tokens issued for admin-created accounts carry `forcePasswordChange: true`.
    Those accounts may only call `POST /auth/change-password` until a new
    password is set; all other protected endpoints return `403` until then.

    ## Roles
    Endpoints marked **admin only** require the `role: "admin"` claim in the
    JWT. A valid token with `role: "user"` returns `403`.
servers:
  - url: http://localhost:8080/api/v1
...
security: []   # default — individual operations opt in via security field
paths:
...
  /fixtures:
    get:
      tags: [fixtures]
      summary: List fixtures
      description: |
        Returns fixtures ordered newest first. Defaults to `status: "approved"`
        when the `status` parameter is omitted, preserving scoreboard behaviour.
        Any authenticated user may request any status value.
      operationId: listFixtures
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: query
          required: false
          description: Filter fixtures that involve this team (either side)
          schema:
            type: integer
        - name: status
          in: query
          required: false
          description: |
            Filter by fixture status. Defaults to `approved` when omitted.
          schema:
            $ref: "#/components/schemas/FixtureStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fixture"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [fixtures]
      summary: Submit a new fixture for admin approval
      description: |
        Creates a fixture with `status: "pending"`. The authenticated user is
        recorded as `submittedBy`. An admin must approve it before it counts
        toward scores.

        `playedAt` defaults to the current server time when omitted.
        `team1Score` and `team2Score` must either both be provided or both be omitted.
      operationId: createFixture
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFixtureRequest"
      responses:
        "201":
          description: Created — fixture is pending admin review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error (same team on both sides, or only one score provided)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /fixtures/{id}:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      tags: [fixtures]
      summary: Get a fixture by ID (any status)
      operationId: getFixture
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [fixtures]
      summary: Update a fixture (result, scores, value)
      description: |
        `team1Score` and `team2Score` must either both be provided or both be omitted.
      operationId: updateFixture
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFixtureRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Validation error (only one score provided)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [fixtures]
      summary: Delete a fixture
      operationId: deleteFixture
      security:
        - BearerAuth: []
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /fixtures/{id}/approve:
    parameters:
      - $ref: "#/components/parameters/id"

    patch:
      tags: [fixtures]
      summary: Approve a pending fixture (admin only)
      description: |
        Transitions the fixture from `pending` to `approved`. The fixture will
        now appear in `GET /fixtures` and contribute to player scores.
        Returns `409` if the fixture is not currently `pending`.
      operationId: approveFixture
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK — returns the updated fixture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Fixture is not in pending state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /fixtures/{id}/reject:
    parameters:
      - $ref: "#/components/parameters/id"

    patch:
      tags: [fixtures]
      summary: Reject a pending fixture (admin only)
      description: |
        Transitions the fixture from `pending` to `rejected`. The record is
        preserved for audit purposes but does not count toward scores.
        Returns `409` if the fixture is not currently `pending`.
      operationId: rejectFixture
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK — returns the updated fixture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Fixture is not in pending state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# ======================================================= components
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT. Obtain a token from `POST /auth/login` or
        `POST /auth/register`. Pass as `Authorization: Bearer <token>`.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request — malformed JSON or failed field validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not authorised (wrong role, or forcePasswordChange is true)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MatchResult:
      type: string
      enum: [team_1, team_2, draw]
    FixtureStatus:
      type: string
      enum: [pending, approved, rejected]
      description: |
        - `pending` — submitted, awaiting admin review
        - `approved` — counted toward player scores
        - `rejected` — reviewed but excluded from scores (record preserved for audit)

    # ---- entities (read models)
...
    Fixture:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        team1Id:
          type: integer
        team2Id:
          type: integer
        result:
          $ref: "#/components/schemas/MatchResult"
        team1Score:
          type: integer
          nullable: true
        team2Score:
          type: integer
          nullable: true
        playedAt:
          type: string
          format: date-time
        value:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/FixtureStatus"
          readOnly: true
        submittedBy:
          type: integer
          nullable: true
          readOnly: true
          description: ID of the user who submitted this fixture
      required: [id, team1Id, team2Id, result, playedAt, value, status]
...
    CreateFixtureRequest:
      type: object
      properties:
        team1Id:
          type: integer
        team2Id:
          type: integer
        result:
          $ref: "#/components/schemas/MatchResult"
        team1Score:
          type: integer
          nullable: true
          minimum: 0
        team2Score:
          type: integer
          nullable: true
          minimum: 0
        playedAt:
          type: string
          format: date-time
          description: Defaults to the current server time when omitted
        value:
          type: integer
          minimum: 0
          default: 0
      required: [team1Id, team2Id, result]
    UpdateFixtureRequest:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/MatchResult"
        team1Score:
          type: integer
          nullable: true
          minimum: 0
        team2Score:
          type: integer
          nullable: true
          minimum: 0
        value:
          type: integer
          minimum: 0
      required: [result]
...
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]